Add files via upload
A WhatsApp Web clone is a web application designed to mimic the functionality of
the popular WhatsApp messenger app on the web.
It is one of those online web development projects where you will build a clone
app allowing users to use WhatsApp on their laptops or desktops without
installing any mobile application. This application typically uses similar user
interfaces and functionality to the official WhatsApp Web application using HTML,
JavaScript, CSS, and WebSockets for real-time communication. It is a great
project to showcase your web development skills as you will have to be careful in
mimicking all features, integrating stickers and emoticons, and facilitating voice
and video calls..